To add methods: checkPhoneNumber(), checkMatch(), formatName(), formatFullName(), formatPhoneNumber()
Method public static boolean checkPhoneNumber(String)
checks the correctness of entering the phone number.
The criterion: presence of 11 digits in the entered line.

Method public static int checkMatch
(String, String[][], int) checks if the first argument
matches any string of the second argument in the column
specified in the third argument? If there are no matches,
it returns -1, otherwise it returns the row's index of the
match.

Method public static String formatName(String) formats an
argument (word) and returns a string with the same symbols
as in the argument, but with the first letter in upper case
and the rest in lower case.

Method public static String formatFullName(String) splits
the argument into separate words and uses the formatName()
method to format as described above. It also formats
hyphenated words so that both parts of such words start
with a capital letter.

Method public static String formatPhoneNumber(String)
formats the argument and returns a string in the format
+7 XXX XXX XX XX.