To add method delRegister()
Method public static String[][] delRegister(String[][],
int) deletes a register by phone number or by name. If the
number of free rows after deleting a register becomes more
than five, then it decreases the size  of the array in
such a way that for registers there are two free rows.